Refactor project property page
This splits the Ivy related UI and logic to its own class to pave
the way for an alternate UI for Maven projects.
